requirements.txt
cycler==0.11.0
defusedxml==0.7.1
fonttools==4.33.3
fpdf2==2.5.4
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
reportlab==3.6.9
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
